% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_parameters.r
\name{make.batch.file}
\alias{make.batch.file}
\title{Make batch file}
\usage{
make.batch.file(
  parameter_list,
  executable_path,
  n_replicates = 1,
  n_cores = 1,
  batch_file_prefix = "batch_file"
)
}
\arguments{
\item{parameter_list}{a list or dataframe with all parameter combinations
the order with which a parameter was added to the list will be
maintained}

\item{executable_path}{the path to the simulation executable (i.e.,
the exe file)}

\item{n_replicates}{the number of replicates you'd like to run
of each simulation}

\item{n_cores}{the number of cores available. If n_cores>1, simulations
can be run in parallel}

\item{batch_file_prefix}{the prefix of the resulting batch file to which
all the output will be written to}
}
\value{
nothing, but a file with the path batch_file_prefix will be
written to disk
}
\description{
This function produces a batch file
in the following format:
/path/to/executable.exe 0.01 1 0 output_file_1.csv
/path/to/executable.exe 0.01 2 0 output_file_2.csv
/path/to/executable.exe 0.02 1 0 output_file_3.csv
etc
}
\examples{
:
#initalize parameter object
parameter_object <- list()
parameter_object$a <- c(0.0,0.01)
parameter_object$b <- 0.1

make_batch_file(parameter_list=parameter_object
                     ,executable_path="./my_simulation.exe")
# which produces a file in the current directory
# called 'batch_file.sh'
# with the contents
# ./my_simulation.exe 0.0 0.1 output_file_1
# ./my_simulation.exe 0.001 0.1 output_file_1
}
