% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_simulation.r
\name{summarize.sims}
\alias{summarize.sims}
\title{Summarizes a collection of simulation files, by producing a
\code{data.frame}
containing parameters and the data from the last generation}
\usage{
summarize.sims(
  simulations_path,
  simulation_file_pattern = "sim_.*",
  parameter_start_pattern = "^var",
  parameter_end_pattern = NA,
  data_start_pattern = "^generation",
  data_end_pattern = "^\\n",
  sep = ";",
  recursive = T
)
}
\arguments{
\item{simulations_path}{the directory in which all the simulations are
collected}

\item{simulation_file_pattern}{a \href{
https://cran.r-project.org/web/packages/stringr/vignettes/regular-expressions.html
}{regular expression} that matches the simulation files}

\item{parameter_start_pattern}{a regular expression that matches the start
of the parameters}

\item{parameter_end_pattern}{a regular expression that matches the end
of the parameters. If parameters are at the end of the file anyway
one can just write \code{NA}}

\item{data_start_pattern}{a regular expression that matches the start
of the data output for each timestep or generation}

\item{data_end_pattern}{a regular expression that matches the end
of the data output for each timestep or generation}

\item{sep}{data separators, e.g., \code{","} or \code{"\t"}}

\item{recursive}{if \code{recursive = TRUE}, the search for simulation files
recurses into subdirectories. If \code{recursive = FALSE}, only the current
directory will be searched without descending into subdirectories}
}
\value{
A \code{data.frame} containing the parameters and last line of
output of each simulation file and the corresponding file name
}
\description{
Summarizes a collection of simulation files, by producing a
\code{data.frame}
containing parameters and the data from the last generation
}
\examples{
# say the current directory "."
# contains the following files:
# ├── file_x.csv
# ├── parameters.txt
# ├── sim_cue_integration_23_06_2020_095246_1.csv
# ├── sim_cue_integration_23_06_2020_095246_10.csv
# ├── sim_cue_integration_23_06_2020_095246_10_dist.csv
# ├── sim_cue_integration_23_06_2020_095246_11.csv
# ├── sim_cue_integration_23_06_2020_095246_11_dist.csv
#
data <- summarize.sims(simulations_path="."
                           ,simulation_file_pattern="sim_.*?\\\\d.csv"
                           )
# collects the output from the files in the data.frame data
# sim_cue_integration_23_06_2020_095246_1.csv
# sim_cue_integration_23_06_2020_095246_10.csv
# sim_cue_integration_23_06_2020_095246_11.csv

# data will look like
str(data)
# 'data.frame':	150 obs. of  3 variables:
# generation: num 50000 500000 500000 ...
# var1: num 1 2 3 4 ...
# x: num 0.33 0.35 0.35...
}
